authorizer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT==2.9.0
httpx==0.27.2
pydantic==2.8.2
